Fetch FIlm Hook
created fetch film hook
